Autogenerate the resolution system
