+++ codegen-sdk-ast-generator/src/lib.rs
@@ -1,3 +1,4 @@
+// Generate the AST and type resolution for a language
 #![feature(extend_one)]
 
 use codegen_sdk_common::{generator::format_code, language::Language};
@@ -5,6 +6,7 @@ use codegen_sdk_cst::CSTDatabase;
 use quote::{ToTokens, quote};
 mod generator;
 mod query;
+mod resolution;
 pub use query::{GROUPS, HasQuery, field::Field, symbol::Symbol};
 mod visitor;
 use syn::parse_quote;
@@ -18,15 +20,23 @@ pub fn generate_ast(language: &Language) -> anyhow::Result<()> {
     use std::collections::BTreeMap;
     use codegen_sdk_resolution::HasFile;
     use codegen_sdk_resolution::Parse;
+    use codegen_sdk_resolution::{ResolveType, Scope};
+    use codegen_sdk_ast::{Definitions as _, References as _};
+    use codegen_sdk_resolution::ResolutionStack;
+    use codegen_sdk_resolution::HasId;
+    use codegen_sdk_resolution::{Dependencies as _};
+
     };
     let ast = generator::generate_ast(language)?;
     let definition_visitor = visitor::generate_visitor(&db, language, "definition");
     let reference_visitor = visitor::generate_visitor(&db, language, "reference");
+    let resolution = resolution::generate_resolution(language);
     let ast: syn::File = parse_quote! {
         #imports
         #ast
         #definition_visitor
         #reference_visitor
+        #(#resolution)*
     };
     let out_dir = std::env::var("OUT_DIR")?;
     let out_file = format!("{}/{}-ast.rs", out_dir, language.name());
