+++ codegen-sdk-resolution/src/references.rs
@@ -14,21 +14,4 @@ pub trait References<
     where
         Self: Sized,
         Scope: Parse<'db>;
-    // {
-    //     // let files = files(db);
-    //     // log::info!(target: "resolution", "Finding references across {:?} files", files.len());
-    //     // let mut results = Vec::new();
-    //     // for input in files {
-    //     //     // if !self.filter(db, &input) {
-    //     //     //     continue;
-    //     //     // }
-    //     //     let file = Scope::parse(db, input.clone());
-    //     //     let dependencies = file.clone().compute_dependencies_query(db);
-    //     //     if let Some(references) = dependencies.get(db, &self.fully_qualified_name(db)) {
-    //     //         results.extend(references.iter().cloned());
-    //     //     }
-    //     // }
-    //     results
-    // }
-    fn filter(&self, db: &'db dyn Db, input: &codegen_sdk_cst::File) -> bool;
 }
